# openapi-crm-springboot.yaml
openapi: 3.0.3
info:
  title: CRM Backend API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /leads:
    get:
      tags: [Leads]
      summary: List leads
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: assignedUserId
          schema: { type: integer, format: int64 }
        - in: query
          name: fromDate
          schema: { type: string, format: date }
        - in: query
          name: toDate
          schema: { type: string, format: date }
        - in: query
          name: leadType
          schema: { type: string }
        - in: query
          name: leadSource
          schema: { type: string }
      responses:
        '200':
          description: Lead list
    post:
      tags: [Leads]
      summary: Create lead (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LeadCreateMultipart'
      responses:
        '201':
          description: Lead created
...
  /opportunities:
    get:
      tags: [Opportunities]
      summary: List opportunities
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Opportunities
    post:
      tags: [Opportunities]
      summary: Create opportunity
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OpportunityCreateMultipart'
      responses:
        '201':
          description: Opportunity created
  /opportunities/{id}:
    get:
      tags: [Opportunities]
      summary: Get opportunity by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Opportunity
    put:
      tags: [Opportunities]
      summary: Update opportunity
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OpportunityCreateMultipart'
      responses:
        '200':
          description: Opportunity updated
    delete:
      tags: [Opportunities]
      summary: Delete opportunity
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Opportunity deleted
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      responses:
        '200':
          description: Projects
    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectCreateMultipart'
      responses:
        '201':
          description: Project created
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Project
    put:
      tags: [Projects]
      summary: Update project
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectCreateMultipart'
      responses:
        '200':
          description: Project updated
    delete:
      tags: [Projects]
      summary: Delete project
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Project deleted
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      responses:
        '200':
          description: Tasks
    post:
      tags: [Tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TaskCreateMultipart'
      responses:
        '201':
          description: Task created
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Task
    put:
      tags: [Tasks]
      summary: Update task
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated
    delete:
      tags: [Tasks]
      summary: Delete task
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Task deleted
...
  /organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      responses:
        '200':
          description: Organizations
    post:
      tags: [Organizations]
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
      responses:
        '201':
          description: Organization created
  /organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Organization
    put:
      tags: [Organizations]
      summary: Update organization
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
      responses:
        '200':
          description: Organization updated
    delete:
      tags: [Organizations]
      summary: Delete organization
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Organization deleted
  /roles:
    get:
      tags: [Roles]
      summary: List roles
      responses:
        '200':
          description: Roles
    post:
      tags: [Roles]
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleName:
                  type: string
      responses:
        '201':
          description: Role created
  /roles/{id}:
    get:
      tags: [Roles]
      summary: Get role by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Role
    put:
      tags: [Roles]
      summary: Update role
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleName:
                  type: string
      responses:
        '200':
          description: Role updated
    delete:
      tags: [Roles]
      summary: Delete role
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Role deleted
  /roles/{id}/permissions:
    get:
      tags: [Roles]
      summary: Get role permissions
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Permissions
    put:
      tags: [Roles]
      summary: Replace role permissions
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Permissions updated
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    LeadCreateMultipart:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        title: { type: string }
        mobileNo: { type: string }
        address: { type: string }
        email: { type: string }
        city: { type: string }
        state: { type: string }
        country: { type: string }
        phoneNo: { type: string }
        organizationName: { type: string }
        website: { type: string }
        industry: { type: string }
        noOfEmployee: { type: integer }
        source: { type: string }
        type: { type: string }
        reason: { type: string }
        status: { type: string }
        designation: { type: string }
        inquiryDate: { type: string, format: date }
        userIdFk: { type: integer, format: int64 }
        uploadDocument: { type: string, format: binary }
        uploadDocument1: { type: string, format: binary }
        uploadDocument2: { type: string, format: binary }
        uploadDocument3: { type: string, format: binary }
...
    OpportunityCreateMultipart:
      type: object
      properties:
        oppName: { type: string }
        oppTitle: { type: string }
        oppStatus: { type: string }
        oppAmount: { type: number }
        oppForcastCloseDate: { type: string, format: date }
        oppActualCloseDate: { type: string, format: date }
        oppDescription: { type: string }
        userIdFk: { type: integer, format: int64 }
        docOpportunity: { type: string, format: binary }
    ProjectCreateMultipart:
      type: object
      properties:
        projectName: { type: string }
        organizationName: { type: string }
        projectStatus: { type: string }
        projectStartDate: { type: string, format: date }
        projectCompletedDate: { type: string, format: date }
        forecastCompletedDate: { type: string, format: date }
        projectDescription: { type: string }
        uploadProjectDoc: { type: string, format: binary }
    TaskCreateMultipart:
      type: object
      properties:
        taskName: { type: string }
        taskAssign:
          type: string
          enum: [Member, Team]
        taskAssignedMemberIds:
          type: array
          items: { type: integer, format: int64 }
        taskAssignedTeamId:
          type: integer
          format: int64
        taskStartDate: { type: string, format: date }
        taskCompletedDate: { type: string, format: date }
        taskDueDate: { type: string, format: date }
        taskRelatedTo: { type: string }
        taskDescription: { type: string }
        taskPriority: { type: string }
        taskPercentageCompleted: { type: integer }
        uploadTask: { type: string, format: binary }
    TaskUpdateRequest:
      type: object
      properties:
        taskName: { type: string }
        taskAssignedTo: { type: integer, format: int64 }
        taskStartDate: { type: string, format: date }
        taskCompletedDate: { type: string, format: date }
        taskDueDate: { type: string, format: date }
        taskRelatedTo: { type: string }
        taskDescription: { type: string }
        taskPriority: { type: string }
        taskPercentageCompleted: { type: integer }
...
    OrganizationRequest:
      type: object
      properties:
        organizationName: { type: string }
        organizationMoblieNo: { type: string }
        organizationEmail: { type: string }
        organizationAddress: { type: string }
        organizationCity: { type: string }
        organizationState: { type: string }
        organizationCountry: { type: string }
        organizationBackground: { type: string }
        organizationOccasion: { type: string }
        organizationPostcode: { type: string }
        organizationOccasionDate: { type: string, format: date }
...